+++ CHANGELOG.rst
@@ -17,3 +17,12 @@ Changes
 
 * rm _set_today in model
 * Add /odk to openrosa url
+
+0.2.0 (2021-09-17)
+------------------
+
+* Remove dependancy on lxml
+* Remove modifications on xml_content field
+* Add admin management
+* Add use of AUTH_USER_MODEL in settings.py
+* Add delete button on submitted forms